Compact Compressor UI (#84116)
## About The Pull Request

Made the tank compressor UI more compact, removing the data on gas
mixes.

## Before:


![image](https://github.com/tgstation/tgstation/assets/3625094/cd857de3-6c92-46f7-9eee-9a4a801421fb)

## After:


![image](https://github.com/tgstation/tgstation/assets/3625094/51f09447-c6e8-4ac2-ac60-c2d4821adaff)

## Why It's Good For The Game

The old UI was overcomplicated, the gas mixes can be analyzed with gas
analyzer.

## Changelog

:cl:
refactor: Compressor UI to TypeScript
qol: Simplified Compressor UI layout
/:cl:
